requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
decorator==5.1.0
future==0.18.2
geocoder==1.38.1
idna==3.2
python-dotenv==0.19.0
ratelim==0.1.6
requests==2.26.0
six==1.16.0
urllib3==1.26.7
